backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.0
SQLAlchemy==2.0.33
psycopg2==2.9.6
python-dotenv==1.0.0
alembic==1.11
pydantic[email]
passlib
bcrypt==3.2.0
pytest==7.4.0
httpx==0.24.1

python-jose==3.3.0

python-multipart==0.0.6
jinja2==3.1.2
aiofiles=23.1.0
